+++ src/frontend/src/css/OrderStatus.css
@@ -19,15 +19,17 @@
 }
 
 .order-title {
-    font-size: 3em;
+    color: var(--frappy-lgreen);
+    font-size: 2em;
     font-weight: 600;
     margin-top: 1em;
+    margin-bottom: 0.5em;
     padding-left: 4em;
     padding-right: 4em;
 }
 
 .status-container {
-    flex-grow: 1;
+    flex-grow: 0;
     display: flex;
     justify-content: center;
     align-items: center;
@@ -56,13 +58,15 @@
 }
 
 .timer-container {
-    flex-grow: 0.6;
+    font-size: 1em;
+    flex-grow: 1;
     display: flex;
     flex-direction: column;
     justify-content: flex-start;
     align-items: center;
 
 }
+
 .wait-text {
     font-size: 2em;
     font-weight: 600;
